+ Arreglado el memory leak en HUD + Quitado un zillion de warnings de generator.

+++ generator/source/include/GenLagoonZone.h
@@ -38,8 +38,8 @@ class GenLagoonZone : public GenZone {
 			Metaball(int X, int Y, float R, int type)
 			{
 				this->type = type;
-				x = X;
-				y = Y;
+				x = (float)X;
+				y = (float)Y;
 				radius = R;
 				xm = ym = 1.0f;
 			}
