+++ generator/source/src/GenForestZone.cpp
@@ -104,7 +104,7 @@ void GenForestZone::placeDungeon()
 		short tries = 0;
 		
 		while (!placed){
-			if (tile < overworld->mapTileMatrix->size() &&
+			if (tile < (int)overworld->mapTileMatrix->size() &&
 				overworld->mapTileMatrix->at(tile)->getZoneNumber() == this->zoneNumber && 
 				overworld->mapTileMatrix->at(tile)->getSolid() > 0 ){
 				if ( !isFrontierNear(tile, range) ){
@@ -221,7 +221,7 @@ bool GenForestZone::isFrontierNear(int iniT, int range){
 		tile = iniTile + i*overworld->getWorldSizeW();
 		for (int j = 0; j < (range*2+1); j++){
 			if ( !frontierFound) 
-				if (tile >= overworld->mapTileMatrix->size() || overworld->mapTileMatrix->at(tile)->getSolid() == 4
+				if (tile >= (int)overworld->mapTileMatrix->size() || overworld->mapTileMatrix->at(tile)->getSolid() == 4
 					|| overworld->mapTileMatrix->at(tile)->getZoneNumber() != overworld->mapTileMatrix->at(iniT)->getZoneNumber() )
 					frontierFound = true;
 			tile++;
@@ -318,7 +318,7 @@ void GenForestZone::placeEntrance(int entrance){
 		for (short col = 0; col<maxCols; col++){
 			tile = iniT + col*overworld->getTileWorldSizeW();
 			for (short row = 0; row<maxRows; row++){
-				if ( tile < overworld->mapTileMatrix->size())
+				if ( tile < (int)overworld->mapTileMatrix->size())
 					overworld->mapTileMatrix->at(tile)->setSolid(1);
 				tile++;
 			}
@@ -333,7 +333,7 @@ void GenForestZone::genGeoDetail()
 {
 	int numScreens = screenList->size();
 	sowSeeds(numScreens-numScreens/5); //Plantamos semillas
-	int numSolids = floor((numScreens*SCREEN_HEIGHT*SCREEN_WIDTH)*0.60);
+	int numSolids = (int)floor((numScreens*SCREEN_HEIGHT*SCREEN_WIDTH)*0.60);
 	waterSeeds(numSolids);  //regamos las semillas para que se reproduzcan
 }
 
@@ -349,7 +349,7 @@ bool GenForestZone::sowSeeds(int numSeeds)
 
 	//plantamos una semilla por pantalla
 	int i = 0;
-	while (i < screenList->size() && sowed < numSeeds)
+	while (i < (int)screenList->size() && sowed < numSeeds)
 	{
 		screenN = screenList->at(rand()%screenList->size())->getScreenNumber();  //cogemos una screen
 
@@ -403,7 +403,7 @@ bool GenForestZone::waterSeeds(int numSolids)
 	int growFactor;
 	int actualSeed = 0;
 	int actualSolids = 0;
-	while(actualSeed < seeds->size() && actualSolids < numSolids)
+	while(actualSeed < (int)seeds->size() && actualSolids < numSolids)
 	{
 		growFactor = rand()%51 + 40;
 		actualSolids = actualSolids + growSeed(seeds->at(actualSeed), growFactor);
